fix(gpu): address round-1 review — VRAM gate default, row_offset ABI, zero-threads guard
Cursor Agent - Grok 4.6 High Fast:
- `SCX_REQUIRE_LARGE_VRAM` was opt-in, so the one test that can observe the
  2^31 kernel overflow could skip on a default harness run and the suite would
  still go green. `_run_scx_gpu_tests.sh` now defaults it on for the same reason
  it already forces SCX_REQUIRE_GPU=1 — the job holds a whole H100 via
  --gres=gpu:1, so the ~10 GiB headroom is a property of the allocation.
  Override with SCX_REQUIRE_LARGE_VRAM=0 on a smaller or shared GPU.
- `gpu_hvg.rs` was named as an uncovered site: its six launchers counted in u64
  but cast the block count to u32 with no grid_dim.x cap check. All six now go
  through `flat_launch_1d`, which makes the crate's 1-D launch geometry uniform
  (15 sites).

gemini-code-assist:
- `row_scale_kernel`'s comment claimed the kernel was "correct on its own terms
  rather than by the caller's grace", but `row_offset` was still an `int`
  PARAMETER — an out-of-range value truncated at the ABI boundary before the
  promoted addition saw it, so the host's i32::MAX guard was still what made it
  safe. Widen the parameter to `long long`, pass i64, and drop the now-pointless
  host cap; the `factors.len()` bounds guard above it is the real check.
- `flat_launch_1d` validated `threads > 0` with a `debug_assert!`, which is
  compiled out in release, leaving `div_ceil` to panic on a bare divide-by-zero.
  It returns a Result and is re-exported from the crate root, so make it a real
  check. New test watched failing against the assert form first.

codex - gpt-5.6-sol:
- The new boundary test used `CudaStream::memcpy_dtov`, deprecated in cudarc in
  favour of `clone_dtoh`. CI's clippy does not lint test targets so it passed,
  but `cargo clippy -p scx-gpu --lib --tests` flagged it. Switched.

PR body corrected in place with visible correction blocks for the two claims
this round proved wrong.

+++ scx-gpu/kernels/normalize_log1p.cu
@@ -93,11 +93,17 @@ extern "C" __global__ void log1p_kernel(
 // factor is supplied. `factors` is a global per-row vector indexed in the
 // source's iteration row order; `row_offset` is this shard's first global
 // row, so one device factor vector serves every shard in a streamed source.
+// `row_offset` is `long long` (review §8.4): `factors` is indexed in the
+// *global* obs row space, so `row_offset + row` is an n_obs-scale quantity even
+// though each shard's `n_rows` is small. Promoting only the addition would have
+// left the value truncating at the parameter boundary instead — the host would
+// still have had to cap `row_offset` at i32::MAX for the kernel to be safe,
+// which is a cap on the obs axis with no reason to exist.
 extern "C" __global__ void row_scale_kernel(
     const long long* __restrict__ indptr,
     float* __restrict__ data,
     int n_rows,
-    int row_offset,
+    long long row_offset,
     const float* __restrict__ factors
 ) {
     int row = blockIdx.x * blockDim.x + threadIdx.x;
@@ -106,12 +112,7 @@ extern "C" __global__ void row_scale_kernel(
     long long start = indptr[row];
     long long end = indptr[row + 1];
 
-    // 64-bit global row id (review §8.4): `factors` is indexed in the *global*
-    // obs row space, so `row_offset + row` is an n_obs-scale quantity even
-    // though each shard's `n_rows` is small. The host already refuses a
-    // row_offset past i32::MAX (`gpu_preprocess.rs`); this makes the kernel
-    // correct on its own terms rather than by the caller's grace.
-    float factor = factors[(long long)row_offset + row];
+    float factor = factors[row_offset + row];
     for (long long i = start; i < end; i++) {
         data[i] *= factor;
     }
